+++ 404.php
@@ -4,14 +4,23 @@
 	<div class="container">
 		<div class="row">
 			<div class="<?=FULLWIDTH?>">
-				<h1><?=__('Oh no!','sherpa')?></h1>
+				<h1><?=__("Oh no!  It's a 404!",'sherpa')?></h1>
 				<h2><?=__("We looked really hard but we couldn't find it!",'sherpa')?></h2>
 				
-				<p><?=__("In the meanwhile, we've found this funny animal video for you to watch.  Enjoy!",'sherpa')?></p>
-				
-				<div class="embed-responsive embed-responsive-16by9">
-    				<iframe src="https://www.youtube.com/embed/EQ1HKCYJM5U" allowfullscreen class="embed-responsive-item"></iframe>
-				</div>
+				<h4><?=__("In the meanwhile, let's see if you were looking for one of these pages",'sherpa')?>:</h4>
+				<ul>
+                <?php
+                    // Make sure to add in the IDs of any pages you want to exclude such as Thank You pages
+                    
+                    $args = array(
+                        'title_li' => '',
+                        'exclude' => ''
+                    );
+                    
+                    wp_list_pages($args);  
+                ?>
+				</ul>				
+
 			</div>
 		</div>
 	</div>